fix cross-compile datalayout by extracting TargetInfo to dependency-free module
